fix: fix click in modal
