+++ src/components/index.ts
@@ -3,5 +3,6 @@ export * from '@/components/Footer'
 export * from '@/components/Header'
 export * from '@/components/Lang'
 export * from '@/components/Layout'
+export * from '@/components/Modal'
 export * from '@/components/Navbar'
 export * from '@/components/Toggle'
